todo-list: tried to fix context
